文件日志和swagger配置

+++ rpp-db/src/main/java/com/byoryn/rppdb/utils/Code2EnumUtil.java
@@ -8,7 +8,7 @@ import com.byoryn.rppdb.base.BaseTypeEnum;
  * 根据typeCode得到对应的枚举对象
  */
 public class Code2EnumUtil {
-    public static BaseTypeEnum codeOf(Class<? extends BaseTypeEnum> enumClass, byte typeCode) {
+    public static BaseTypeEnum codeOf(Class<? extends BaseTypeEnum> enumClass, int typeCode) {
         BaseTypeEnum[] enumConstants = enumClass.getEnumConstants();
         for (BaseTypeEnum typeEnum : enumConstants) {
             if (typeEnum.getTypeCode() == typeCode)
@@ -16,4 +16,5 @@ public class Code2EnumUtil {
         }
         return null;
     }
+
 }
